refactor(server): Move DefaultContext to zally-core #560
