+++ server/src/main/java/de/zalando/zally/configuration/ContextFactoryConfiguration.kt
@@ -1,0 +1,19 @@
+package de.zalando.zally.configuration
+
+import de.zalando.zally.core.DefaultContextFactory
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import java.util.regex.Pattern
+
+@Configuration
+class ContextFactoryConfiguration {
+
+    @Value("\${zally.propagateAuthorizationUrls:}")
+    private lateinit var propagateAuthorizationUrls: List<Pattern>
+
+    @Bean
+    fun contextFactory(): DefaultContextFactory {
+        return DefaultContextFactory()
+    }
+}
