+++ server/src/main/java/de/zalando/zally/configuration/RulesValidatorConfiguration.kt
@@ -1,9 +1,9 @@
 package de.zalando.zally.configuration
 
+import de.zalando.zally.core.DefaultContextFactory
 import de.zalando.zally.rule.ApiValidator
 import de.zalando.zally.rule.CompositeRulesValidator
 import de.zalando.zally.rule.ContextRulesValidator
-import de.zalando.zally.rule.DefaultContextFactory
 import de.zalando.zally.rule.JsonRulesValidator
 import de.zalando.zally.rule.RulesManager
 import org.springframework.beans.factory.annotation.Autowired
