+++ server/src/test/java/de/zalando/zally/rule/RulesValidatorTest.kt
@@ -3,6 +3,7 @@ package de.zalando.zally.rule
 import com.fasterxml.jackson.core.JsonPointer
 import com.fasterxml.jackson.databind.JsonNode
 import com.typesafe.config.ConfigFactory
+import de.zalando.zally.core.ContentParseResult
 import de.zalando.zally.core.EMPTY_JSON_POINTER
 import de.zalando.zally.rule.api.Check
 import de.zalando.zally.rule.api.Rule
