+++ server/src/test/java/de/zalando/zally/rule/zalando/ApiAudienceRuleTest.kt
@@ -1,9 +1,9 @@
 package de.zalando.zally.rule.zalando
 
 import de.zalando.zally.core.rulesConfig
-import de.zalando.zally.rule.DefaultContextFactory
-import de.zalando.zally.rule.api.Context
 import de.zalando.zally.test.ZallyAssertions.assertThat
+import de.zalando.zally.core.DefaultContextFactory
+import de.zalando.zally.rule.api.Context
 import org.junit.Test
 
 class ApiAudienceRuleTest {
