+++ server/src/test/java/de/zalando/zally/rule/zalando/FunctionalNamingForHostnamesRuleTest.kt
@@ -1,6 +1,6 @@
 package de.zalando.zally.rule.zalando
 
-import de.zalando.zally.rule.DefaultContextFactory
+import de.zalando.zally.core.DefaultContextFactory
 import de.zalando.zally.rule.api.Context
 import org.assertj.core.api.Assertions.assertThat
 import org.intellij.lang.annotations.Language
