+++ server/src/test/java/de/zalando/zally/rule/zalando/LimitNumberOfSubResourcesRuleTest.kt
@@ -1,7 +1,7 @@
 package de.zalando.zally.rule.zalando
 
-import de.zalando.zally.rule.DefaultContextFactory
 import de.zalando.zally.core.rulesConfig
+import de.zalando.zally.core.DefaultContextFactory
 import org.assertj.core.api.Assertions.assertThat
 import org.intellij.lang.annotations.Language
 import org.junit.Test
