+++ server/zally-core/src/main/kotlin/de/zalando/zally/core/DefaultContext.kt
@@ -1,10 +1,9 @@
-package de.zalando.zally.rule
+package de.zalando.zally.core
 
 import com.fasterxml.jackson.core.JsonPointer
-import de.zalando.zally.core.toSwaggerJsonPointer
+import de.zalando.zally.core.ast.ReverseAst
 import de.zalando.zally.rule.api.Context
 import de.zalando.zally.rule.api.Violation
-import de.zalando.zally.core.ast.ReverseAst
 import io.swagger.models.Swagger
 import io.swagger.v3.oas.models.OpenAPI
 import io.swagger.v3.oas.models.Operation
